pyproject.toml
[project]
name = "langchain-case-py"
version = "0.1.0"
description = "LangChain Python 案例集 - 对应 TypeScript 版本的 langchain-case 项目"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain>=1.2.0",
    "langchain-openai>=0.3.0",
    "langgraph>=0.4.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.0.0",
    "ruff>=0.3.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100

[tool.ruff]
line-length = 100
